Committing after implementing jsonwebtoken

+++ Medicatch_Backend/controllers/articlesControllers.js
@@ -1,5 +1,8 @@
 const articleModel = require("../models/Articles");
 exports.createArticle = async (req, res) => {
+  if (req.user.role !== "admin") {
+    return res.status(403).json({ message: "Access denied. Admins only." });
+  }
   try {
     const { title, content } = req.body;
     const article = new articleModel({ title, content });
@@ -41,6 +44,9 @@ exports.getIndividualArticle = async (req, res) => {
 };
 
 exports.updateAnArticle = async (req, res) => {
+  if (req.user.role !== "admin") {
+    return res.status(403).json({ message: "Access denied. Admins only." });
+  }
   try {
     const { id } = req.params;
     const { title, content } = req.body;
@@ -62,6 +68,9 @@ exports.updateAnArticle = async (req, res) => {
 };
 
 exports.deleteAnArticle = async (req, res) => {
+  if (req.user.role !== "admin") {
+    return res.status(403).json({ message: "Access denied. Admins only." });
+  }
   try {
     const { id } = req.params;
     const article = await articleModel.findByIdAndDelete(id);
@@ -74,4 +83,4 @@ exports.deleteAnArticle = async (req, res) => {
   } catch (error) {
     res.status(500).json({ message: "Error deleting article", error });
   }
-}
+};
